refactor(video): :recycle: Splits SDL backend into Windowing/Rendering folders and trims IRenderer's DrawRect API
Moves SDLWindow/SDLRenderer into Graphics/Windowing/SDL and
Graphics/Rendering/SDLRenderer respectively, fixes the relative includes
broken by the move, and drops IRenderer's redundant float-args DrawRect
overload in favor of the single math::Rect-based one. Updates the
moving_box example and Video interface tests to match.
