+++ examples/moving_box/MovingBox.cpp
@@ -14,8 +14,8 @@ void MovingBox::Update(float inDeltaTime)
 
 void MovingBox::Render(asge::graphics::IRenderer &inRender)
 {
-    inRender.DrawRect( 
-        m_Position.x(), m_Position.y(), m_Dimension.x(), m_Dimension.y(),
+    inRender.DrawRect(
+        asge::math::Rect{ m_Position.x(), m_Position.y(), m_Dimension.x(), m_Dimension.y() },
         {255, 0, 0, 255}
     );
 }
