+++ src/ASGE/CMakeLists.txt
@@ -9,8 +9,8 @@ add_library(ASGE STATIC
     # Video and Graphic portion with Renderers and others
     Video/VideoSystem.cpp
     Video/Graphics/GraphicsFactory.cpp
-    Video/Graphics/SDL/SDLRenderer.cpp
-    Video/Graphics/SDL/SDLWindow.cpp
+    Video/Graphics/Rendering/SDLRenderer/SDLRenderer.cpp
+    Video/Graphics/Windowing/SDL/SDLWindow.cpp
 
     # Engine Logger
     Core/Logger/Logger.cpp
