+++ src/ASGE/Video/Graphics/Rendering/SDLRenderer/SDLRenderer.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../Renderer.hpp"
+#include "../../Renderer.hpp"
 #include <string>
 #include <SDL3/SDL.h>
 
@@ -28,8 +28,6 @@ public:
 
     void Clear(RGBA_Color const& inColor) const override;
     void DrawRect(math::Rect const& inRect, RGBA_Color const& inColor) const override;
-    void DrawRect(float x, float y, float w, float h,
-        RGBA_Color const& inColor) const override;
 
     void Present() const override;
     bool IsValid() const override;
